Ensure consistent group ordering
We starting adding Overview cards from custom metrics and that caused
the metric sections to appear in random places. This seems to happen
because `create-dashboard` used `(partition-by :group)` assuming that
all the cards were continguous groups. This is no longer true.

So I group by group name and then pull the keys out according to the
group score in `ordered-group-by-seq`.