[MLv2] Handle removing multiple dependent joins
`remove-join` was modified to handle dependent joins from subsequent
stages in #34416. However, when a stage had more than two dependent
joins the `m/index-by` call to find the stage number was invalid because
some of the other joins have been removed at that point. The most
straightforward path seemed to simply pass the stage-number of the
clause being removed to `remove-clause` instead of trying to find
something that may have been modified.