Add the ability to specify `:values` for a setting
This allows us to specify `:values` for a given setting.

To design this, I looked at the existing implementation for settings with `keyword` types.

There are a few different patterns for these, and unfortunately they follow quite a few different patterns in terms of
how to validate that a value is a member of the expected set.

Variously, they:

- validate when the setting is read or written, throw an exception if invalid.

- do not validate the setting anywhere (presumably the other code is written to behave sanely in the presence of an
invalid value)

- validate the keyword when it's written, but just set a default when reading.

- validate the keyword when it's written, return whatever is present when reading

This puts us in a bit of an awkward situation:

- if we validate and throw exceptions on read, we may break existing users.

- if we validate on write and warn on read, we need to watch out for cases like `user-visiblity`: the default value is
`:all`, but the fallback from invalid values is `:none`. If someone has a typo for the `MB_USER_VISIBILITY` env var,
changing the logic to fall back to the default (`all`) would unexpectedly change the value of the setting (from `none`
to `all`).

I *think* the most reasonable behavior here is to throw an exception when an invalid value is written, and warn (via
the logs) when an invalid value is read, while returning the default value.

In the next commit, I'll add `:values` to existing `:keyword`-typed Settings in those cases where it makes sense. As
mentioned above, I'll exclude (and add a comment to) the `user-visibility` setting, to prevent an unexpected change
for users who might have set the env var to an invalid value.

One final note: I changed the method implementation of `get-value-of` for the `:keyword` type to lower-case its input
before turning it into a keyword - so setting an env var like `MB_SETTING_VALUE=STRICT` or `MB_SETTING_VALUE=StRiCt`
will both turn into `(= (setting-value) :strict)`. I wanted to do this to allow us to use `:values` for
`SESSION_COOKIE_SAMESITE`, where we've allowed (and had tests for) uppercase/mixed case env vars.