Refactor `defsetting`
Previously, we relied on the functions `get-value-of-type` and `set-value-of-type`. However, I found this confusing,
because the logic for *retrieving the raw value from the system* is (almost) always exactly the same. (See below for
the one exception.) This caused pain when I wanted to define a list of allowed values for a given setting, because
when setting the value, we're dealing with a possible un-parsed value (e.g. `"foo"` instead of `:foo` for a
keyword-typed setting). But because the existing code tied parsing the value to setting or fetching the value, there
was no convenient way to *just parse it*.

So this code rips out the concept of serializing/deserializing a value into a given type and separates it from the
concept of saving or retrieving a setting.

I defined a couple new multimethods:

- `->string` and `string->`, which take a type and a value and convert the type to or from a string, respectively

- `parsed?`, which takes a type and a value and returns `true` if the value is a parsed instance of that type

A few new/modified functions:

- I removed the 3-arity version of `get-raw-value`. That version did not get a raw value. It took functions that
parsed a value and returned a parsed value!

- I also ensured that `get-raw-value` always returned the raw (string) value. Previously, it would sometimes return a
string (raw value) and sometimes return the parsed version of the value (e.g. when it obtained it from the default).
Actually, it sometimes wouldn't even do that - if you asked for a `string` raw value and the default was not a string,
it would just return `nil`!

- I replaced that with `get-parsed-value`. `get-parsed-value` has a 3-arity version, which behaves like
`get-raw-value` used to (it takes a parsing function and a pred-function and uses them to return the parsed version of
the raw value). It also has a single-arity version, which takes a setting name and parses the raw value appropriately
for that setting.

- I deleted the `get-value-of-type` and `set-value-of-type!` multimethods, although I reimplemented them as simple
functions that call `get-parsed-value` and `set-raw-value!`, respectively.

The `setting/get-value-of-type` function was used for two purposes:

- getting `:string` values of settings, in cases where we only want strings. We can replace this with a call to
`get-raw-value`.

- getting the parsed value of a setting. In every case I found, we wanted to parse the value according to the
setting's type - that is, a `:json` setting should get parsed as JSON, a `:boolean` setting should get parsed as a
boolean, etc. These cases can just be replaced with a call to `setting/get-parsed-value`, which uses the `:type` of
the setting to parse it correctly.

The `setting/set-value-of-type!` function was only used in tests. I replaced it with `setting/set-parsed-value!`,
which has essentially the same behavior (just delegating the serialization logic to the type of the setting).

I did notice a possible existing bug in the implementation of `get-value-of-type` for `:string`s - when we retrieve a
setting with `get-value-of-type :string`, we assume that the retrieved value is already a string (the `parse-fn` is
just `identity`). `get-raw-value`'s docs suggest that the type returned *may or may not* be a string - I haven't dug
into the details and wasn't able to find a case where the returned type was not a string, and left the existing
behavior alone.