Fixing cell queries with a temporal unit of year (#35175)
Fixing situations where a cell queries with a temporal unit of year produced nulls in description strings like:

 "A closer look at number of Orders where null of Created At is 2025"

The simple fix was just to add `:year` to the `unit-name` function in
`metabase.automagic-dashboards.names`. In addition to adding a unit
test for this case, several other temporal units were added to the test.

This fixes https://github.com/metabase/metabase/issues/35170.