fix arabic number inconsistency (#34633)
Closes https://github.com/metabase/metabase/issues/34271

### Description

In Arabic locales (`ar` and `ar-sa`), we were displaying Arabic numerals for translated dates, but in the rest of the app we used Western numerals. This PR makes it so that for  dates we also use Western numerals.

Ideally we would have an option for which kind of numeral to use, but it's a sufficient solution to always use Western numerals. Even in Arabic-speaking countries where Arabic numerals are widely used, Western numerals are commonly seen on digital interfaces, so users should not be bothered by it.

### How to verify

1. Go to the Products model.
2. See "Created At" field in English.
3. Go to account settings, change language to "Arabic", then go back to Products model.
4. "Created At" field should now be translated, but using Western numerals.
5. Repeat the same for "Arabic (Saudi Arabia)" locale, and using admin site locale setting.
6. Go back to English, in Products model should see English dates again.

### Demo

https://github.com/metabase/metabase/assets/37751258/826375a2-1f85-4a3d-9abf-fab2e26dd0c0

### Checklist

- [x] Tests have been added/updated to cover changes in this PR