Handle moving dashboards to a different collection type (#35138)
* Fix typo

* Hides public collections in DashboardPicker

When selecting a dashboard for a question in personal collection

* Fix not calling `useCollectionQuery` unnecessarily

Eventhough in this case it won't make any difference, because
`useCollectionQuery` is cached, we still shouldn't do that.

* Handle recently viewed dashboard

* Refactor default import and fix respective tests

* Refactor moving files closer to their usage

* Hide create a new dashboard option

When the question is in a personal collection and is opening public
collection in the dashboard picker. This only happens when users are
opening the root collection as other public collections won't be
visible.

* Hide items in ItemPicker

When the question is in personal collection and users are opening a
public collection. This only happens when users are opening the root
collection because other public collections are not visible.

* Make the boolean more readable

* Remove unused export

* Fix failed unit tests

* Avoid circular dependency

* Simplify dashboard collection type

* Simplify selectors

* Address review: improves hook name

* Review: Avoid using `ComponentProps`

* Review: Avoid using `useEffect` to call callback

* Fix dashboard picker for private questions not searching

* Review: convert a hook without hook usage to function

* Negate the boolean to positive name to make it easier to read

* Remove unrelated import and matches card to dashboard API types

* Handle moving dashboards to a different collection type

Move between personal/public collections