Stop publishing `:event/card-upate` on card (un)archive (#35065)
I noticed this issue when attempting to stop publishing card-update events when a card is pinned.

There are two issues with the existing code:

- first, the `publish-card-update!` function was called with the *new* (updated) card, but the code checked to see
whether the passed card's `:archived` status and the new value of `archived?` were different. So if you asked to
archive a card, the card would be updated to be `:archived true`, and then `(and archived? (not (:archived card)))`
would return `false`. Therefore, we'd *always* fall through to firing an `:event/card-update` event.

- second, if we *had* successfully fired an `:event/card-archive` or `:event/card-unarchive` events, it would have
thrown an exception, because those events do not derive from `:metabase/event`.

To be clear, I'm not entirely sure what the *right* behavior is here. There are a few options:

- we could make the behavior align with the apparent intent of the code, and fire `:event/card-archive` and
`:event/card-unarchive` events when the card is archived or unarchived. However, nothing actually subscribes to these
events.

- a modification of the above: we can just skip firing *any* event when a card is archived or unarchived. Nothing
listens to `:event/card-(un)archive` anyway.

- we could make the code align more clearly with the current behavior. Currently we fire an `:event/card-update` for
every change, including archival. We could just delete the `publish-card-update!` function and always publish that
event.

My initial impulse was to go with the second option (skip firing an event when a card is archived or unarchived) but
after talking with Noah, he suggested that publishing `:event/card-update` would actually be useful for the audit log.
So we'll do that!