Preserve join aliases in dashcard queries (fixes 25278 & 32557) (#35267)
* Preserve join aliases in dashcard queries (fixes 25278 & 32557)

In `query-based-analysis` the final stage was to `maybe-enrich-joins`
to add in joins defined in the query being x-rayed into the dashcard
query. In some cases (e.g. cell queries), the trigger to detect this
addition isn't known as the filters are added downstream.

This PR always adds in the join aliases if they are present since
they should be harmless if not needed and you can't know if they are
needed downstream as the code is currently designed.

The tests assert not only that the cards are created, but that the
filter is present and executable (these previously errored out).

* lint fix

* Squashing `splice-in` into `preserve-joins` and adding some better test criteria for `generate-dashboard-pipeline-test`.