Fixing x-rays with custom columns (#16680)
Queries with custom columns were broken for two reasons:
- When names were computed for display, the humanization code expected fields
- The generated dashcard queries did not preserve the base query's custom expression definitions

For the first point, the code in `metabase.automagic-dashboards.names` was generalized for both fields and expressions.

For the second, the logic that was formerly `preserve-joins` in `metabase.automagic-dashboards.core` has been generalized to `preserve-entity-element` and applied to both joins and expressions.