Only record card views for :card-read events
This change alters when card events trigger view logs, namely, they'll only trigger when :card-read events occur.

Prior, it was using :card-query. I don't entirely know why, I think that's a holdover from the older
view-log/recent-views implementation.

The shape of the view-log schema has also changed so the record-view! function is adjusted to reflect this.

Before, there was a `(complement #{:dashboard :collection})` filter to prevent all pinned questions from being
counted. Now, we DO want pinned card views to count, but only when they have their visualization turned on. Thus, we
can remove `:collection` from the filter.

It is notable that switching from `:card-query` to `:card-read` for the topics that get recorded solved a double
record issue: 2 :card-query events get sent when a user decides to pin a card. Why this happens is TBD (and should be
fixed because the audit-log does get double written here).

Out of scope for this PR, but the audit-log is still recording :card-update, :card-read, :card-query (x2)... so that
log can grow quite a bit still.