Moving x-ray checks to backend for 15655 to reduce flakes
The following x-ray e2e tests are top contributors to x-ray flakes at 6.94% and 6.62%, respectively:
- "X-RAY" should work on a nested question made from base native question (metabase#15655)
- "COMPARE TO THE REST" should work on a nested question made from base native question (metabase#15655)

This change moves the x-ray dashboard generation aspects test load to a unit test and just checks that the insight buttons are present and clickable.

Note that there is still some scenario creation type code in the fixture that could potentially be simplified with some FE help.