Fixing dropdown filter value GET #32126 (#35439)
* Fixing dropdown filter value GET #32126

When GET "/:id/params/:param-key/values" is invoked, the endpoint looks
for valid values to populate for the filter dropdowns using the resolved
field-ids for the underlying item in the dashboard. In some cases, such
as nested queries, the field is not referenced by id, but using the field name.

This PR adds a branch that will look for fields by field ref
(e.g. `[:field "SOURCE" {:base-type :type/Text}]`) if the field cannot be
resolved by id.

Fixes #32126

* lint fix