[MLv2] New, skipped test for removing a breakout field
The original issue pointed to by #22563 is fixed by the overhaul of the
column selection. But trying to reproduce this for the test showed that
the FE doesn't really let you remove a breakout column. It will hide
them, but if you click "Add/Remove columns" they're greyed out.

From an MBQL-Lib POV, though, `(lib/remove-field some-breakout-column)`
succeeds, but the resulting query is broken, and throws if you eg.
`(lib/returned-columns query)`. We should either:
- fix returned/visibile columns, or
- prevent `remove-field` from breaking the query, or
- make `remove-field` on a breakout column remove the breakout?

The test is skipped, but should be passing if we go with fix 1 above.