Fix slow back button x-ray flakes (#35387)
* Attempting to fix x-ray back timeout issues with smaller card count

It looks like the flakes on https://www.deploysentinel.com/ci/analysis?tab=flake for "dashboard back navigation should display a back to the dashboard button in table x-ray dashboards" and "dashboard back navigation should display a back to the dashboard button in model x-ray dashboards" frequently fail at the `cy.wait("@dataset")` stage. I'm guessing this is because the dashboard code is so slow. This PR limits the number of cards produce to `MAX_CARDS` (5) in order to speed up card generation and help with the flakes.

* Adding a timeout to slow x-rays for back button. Hopefully, the small number of max cards makes this unnecessary, though.

* Bumped timeout to 15s instead of 5s. Looks like it was already 5. Again, hopefully the change to minimized the number of generated cards will fix this flake and not take as long.

* Fixing the `show` query param for x-rays

`show` in `core` can be either `:all` or an integer. However, the API layer only exposes the `:all` or `nil` options. This PR exposes the integer path as well.

Note that this is incremental to get flakes working faster. I'll still need to add API tests for this.

* Adding API unit tests to `metabase.api.automagic-dashboards-test` for `show` cases.

* Fixed deftest name to be unique.

* Fixing reference to new card in flaky test