[MLv2] Fix NULL FK values in the QP and drill-thru (#35567)
The QP returns "large int" IDs like FKs as string, since JS doubles only
guarantee 51 bits of mantissa and a 64-bit ID might get rounded.
However NULL IDs were getting returned as `""` empty strings.

This fixes the QP middleware to return `nil` (JS `null`).

In addition, the drill-thru logic in the FE and MLv2 was trying to
filter incorrectly with a NULL value. This PR prevents the FK filter
drill from returning anything in that case.

I filed #35561 to track the fact that would be useful to drill thru to
all rows with a NULL for some FK, but it's low priority.

Fixes #13957.