Adding `publish-event-async!`
The evidence for the slow loads of dashboards such as https://stats.metabase.com/dashboard/2117
as described in https://github.com/metabase/metabase/issues/33499
point to long synchronous calls in `metabase.events/publish-event!`.

This change does the following:

- Adds a span to `publish-event!` for instrumentation
- Adds an async version of the function
- Changes the GET api/dashboard/:id call to use the async function

Based on the discussion in [this slack thread](https://metaboat.slack.com/archives/CKZEMT1MJ/p1697604232592589)
we may not want to retain the async call ATM, but we could use it as
a stopgap for now.

As an aside, we might want to introduce a queue such as activemq if
we really need asynchronous behavior.