Fixing aggregate dropdowns in dashboard filters
The `values-from-card-query` constructs a new query to determine
enumerable values for filter drop-downs in dashboards. It does this
by generating a simple query from the passed in query with the
filtered value as a query filter. This query didn't know about
expressions which the filter might use. The fix was to just add
in the expressions from the base query.

Fixes #12985