Adding additional instrumentation to `publish-event!`
The `publish-events!` defmethod has quite a few code branches and we're trying to figure out if one of these is causing the slowness described in #33499 as well as stats dashboard [2117](https://stats.metabase.com/dashboard/2117). This PR should provide additional information as to where the delays might be occurring.

It adds the following spans/changes:
- An `:events/initialized` attribute at the top level of `publish-event!`
- `record-view!` span
- `record-event!` span
- `update-users-recent-views!` span

If we merge this PR we can decide if these changes are worthwhile and revert if we don't like the additional layer of wrapping.