Fixing bad performance due to massive pprint statements
Spans on stats for massive models are very long, on the order of 4s.

The bad span seems to encompass a small amount of logging and assertion
logic, including a with-out-str pprint, which I've found to be
notoriously slow for large objects. This PR slims down the logged object.

If this is the culprit we should see a substantial performance improvement.