Fix drop-excess-keys
There are two changes here:

- first, call `load-xform-basics` at the *end* of `load-xform` for Collections instead of the beginning. Collections
have a field, `parent_id`, that is translated internally into a `location`. When `load-xform-basics` removes keys that
aren't columns in the database, we remove `parent_id` and the Collection ends up with a null location. Note that the
docstring for `load-xform-basics` previously suggested that it should be called at the *beginning* of any custom
implementation of `load-xform` - but there's really no need for that, as long as we do call it. I updated the
docstring accordingly.

- second, add a new multimethod, `serdes/ingested-model-columns`. The default implementation uses the same method
we've been using - looks in the application DB and returns the set of keys that correspond to the columns there.
However, the multimethod can be overridden in cases like "Action", where the model corresponds to more than one table
in the database (e.g. an Action is actually a row in both the `action` table and one of the `query_action`,
`http_action`, or `implicit_action` tables).