fix tests for loading with extra keys
There were two problems with the existing code:

- running the test left the database in a state that caused subsequent tests to fail :sob:

- running the test _without_ first running `dev/start!` would fail

This led to ... extreme confusion, as I'd try to rerun other failing tests and they'd pass, then try running these
tests and they'd fail... or sometimes pass, depending on if I'd run `start!` from my REPL or not... and then try
running the original failing tests again and they'd fail. Just a lot of degrees of freedom, leading to confusion.

The new test starts with a completely blank database, sets it up, and ensures that we can run
`ensure-audit-db-installed!` without error.