[MLv2] Return query, stage index and column for `column-filter` drill (#35878)
If this drill is targeting an aggregation column, the query needs an
extra stage added. To enable the custom filtering UI, this (possibly 
updated) query, stage number, and the correct column
*in that new stage* are returned as part of the opaque `column-filter` drill.

`lib.js/column-filter-drill-details` is added to hand those details
to the FE as a JS object containing opaque CLJS values.