Fix `data.cols` metadata with a double JOIN in a nested query
The source query metadata gets keyed by just the `:field` ref without
options, which is not unique when a table is included twice.

This change disambiguates the source metadata by `:source_alias`, so
that the correct metadata is copied to the outer query.

The most visible change is that `data.cols[i].source_alias` is now set
correctly, rather than all duplicates being set to the `:source_alias`
of whichever happens to be last in the source query.