[MLv2] Include query, stage and column for quick filter drills
See #35878 which did the same for `column-filter` drills.

This reuses the logic from that PR to include the same information on
`quick-filter` drills as well. Most quick filters are a single click,
but there are a few special cases that are more involved and require
showing a complete filtering UI.

This renames `lib.js/column-filter-drill-details` to
`lib.js/filter-drill-details` as well.