Hydrating POST /api/dashboard results to = GET
This is the first stage of "PUT and POST and endpoint should return data just like calling GET #34828" in which a POST to create a dashboard also hydrates the details needed to return the identical results as if it were a GET.

This enables a single API call vs. two (a POST and and immediate GET).