Renaming fns and reducing root collection hydration for cards
Rebased on master and merged with @calherries `metabase.api.card` changes.

Renamed `hydrate-for-frontend` to `hydrate-card-details` and added a span for observability.

Removed `collection.root/hydrate-root-collection` which makes cards returned from POST and PUT not identical to GET, but only because the `:collection` key is missing, which should already be present on the FE in these situations. Tests were updated to reflect this.