Update dashboard and card api POST and PUT endpoints to return as GET (#36205)
* Adding stub for POST+GET same result

This commit just stubs in the `post-same-as-get-test` test in `metabase.api.dashboard-test`.

* Hydrating POST /api/dashboard results to = GET

This is the first stage of "PUT and POST and endpoint should return data just like calling GET #34828" in which a POST to create a dashboard also hydrates the details needed to return the identical results as if it were a GET.

This enables a single API call vs. two (a POST and and immediate GET).

* Adding updated value to PUT /api/dashboard response

This change adds the same hydration logic to PUT responses as POST responses.

Also added a span around `update-dashboard` to have parity with `get-dashboard`.

* Changed span name for consistency

* Updating create-dashboard-test to include GET comparison

To be consistent with the PUT test this just combines two tests for dashboard creation into one.

* POST/PUT api/card now returns same value as GET

The primary change is using the same hydration logic across the board.

This ensures the same object shape is returned for all operations an removes the need for a GET immediately after a POST or PUT when a create or update action occurs.

* Removing GET on Dashboard Save

When a dashboard is updated via PUT, the succeeding GET operation is no longer needed and has been removed.

* Restored fetch unrelated to dashboard creation.

* Adding =? to `create-a-card` test.

* Updated testing forms to reflect issue being addressed (#34828).

* Reduced expected e2e load count for 21556 test

This e2e test was example based and expected a PUT + 3 additional GETs. With the new code you only need to do 2 GETs since the PUT result is used.

* Renaming fns and reducing root collection hydration for cards

Rebased on master and merged with @calherries `metabase.api.card` changes.

Renamed `hydrate-for-frontend` to `hydrate-card-details` and added a span for observability.

Removed `collection.root/hydrate-root-collection` which makes cards returned from POST and PUT not identical to GET, but only because the `:collection` key is missing, which should already be present on the FE in these situations. Tests were updated to reflect this.

* Reverting FE changes

The FE changes for this are not a subtle as simply removing a follow on GET. It makes more sense to let the BE changes get merged then the the FE team work on leveraging those changes.