Cleaner hydration instrmentation
The goal of this PR is to add instrumentation to `t2/hydrate` without manually adding pervasive spans everywhere in the code.

This PR consists of two ideas to be considered:

This is a small change that adds tracing in a more global fashion to `hydrate-with-strategy`. Unfortunately, this doesn't really capture the top-level `hydrate` calls as we were hoping for. It does provide some additional insights, but not the kind of bulk spans that tell you where you're spending most of your time.

This change is done in `metabase.db` as:

```clojure
(methodical/defmethod t2.hydrate/hydrate-with-strategy :around :default
  "Add OpenTelemetry spans around calls to hydrate-with-strategy."
  [model strategy k instances]
  (span/with-span!
    {:name       "hydrate-with-strategy"
     :attributes {:hydrate/key k}}
    (next-method model strategy k instances)))
```

The idea is simply to add a macro (tentatively in `metabase.api.telemetry`) around `hydrate` and then just use the aliased `te/hydrate` exactly as you would `t2/hydrate`. A few nice things about this approach:

- Easy to implement -- Just change the import
- Reflects the right calling location in instrumentation
- We can modify the macro if desired to dynamically add a different span name. Currently, it's just "hydrate". We could do "hydrate-${entity-type}", for example.
- It is bespoke, but trivial to add or remove. Find and replace the require alias in api nses we want to add spans to or find and replace to remove them.