fix flaky audit tests
somehow (and _only_ in cloverage running) we seem to end up with the
audit collection but no cards in it. Then all of the tests are failing
which check permissions.

```
ERROR in (cannot-edit-audit-content-dashboards-over-api) (metabase/util.cljc:114)
expected: (= "You don't have permissions to do that." (mt/user-http-request :rasta :put 403 (str "dashboard/" (u/the-id dashboard)) {:name "My new title"}))
  actual: java.lang.Exception: Not something with an ID: nil
 at metabase.util$error.invokeStatic (metabase/util.cljc:114)
```

This is failing

```clojure
(deftest cannot-edit-audit-content-dashboards-over-api
  (install-audit-db-if-needed!)
  (let [dashboard (t2/select-one :model/Dashboard :collection_id (:id (perms/default-audit-collection)))]
    (is (= "You don't have permissions to do that."
                                                                   ;; this dashboard is nil
                                                                   ;; and u/the-id errors
           (mt/user-http-request :rasta :put 403 (str "dashboard/" (u/the-id dashboard)) {:name "My new title"})))))
```

So let's ensure that all of the cards, dashboards, and collection exist,
otherwise install them. Previously it was just checking that the
collection existed, and for some reason it was empty