Migrate PulseCards include_csv = true when the card has :display table
Why? This is so that we can eliminate the behaviour of csv attachments being included in emails even when a user
doesn't have them seleted in the dashboard subscription pane. This is confusing behaviour and undesirable in many
cases.

The likely better behaviour is that attachments for any card are only included if explicitly asked for by a user. But,
in order to change the app to behave this way, we should first migrate things so that any user who relies on this old
behavior will still see the attachments they have come to expect.

This PR is still draft and some things must happen still:

 - [ ] the migration is incomplete! At the moment, this will include .csv attachments even for tables with n-rows <
 10. This migration would cause users to see more .csv attachments than before. So, the migration should only set
 include_csv = true when the row count of the results is greater than 10.
 - [ ] the migration # is NOT claimed in the slack channel yet. This should be claimed as soon as the migration
 approach is reviewed to work and an appropriate issue for this change is linked.